+++ railties/test/commands/dbconsole_test.rb
@@ -200,6 +200,49 @@ class Rails::DBConsoleTest < ActiveSupport::TestCase
     assert_match(/Unknown command-line client for db/, output)
   end
 
+  def test_primary_is_automatically_picked_with_3_level_configuration
+    sample_config = {
+      "test" => {
+        "primary" => {
+          "adapter" => "postgresql"
+        }
+      }
+    }
+
+    app_db_config(sample_config) do
+      assert_equal "postgresql", Rails::DBConsole.new.config["adapter"]
+    end
+  end
+
+  def test_specifying_a_custom_connection_and_environment
+    stub_available_environments(["development"]) do
+      dbconsole = parse_arguments(["-c", "custom", "-e", "development"])
+
+      assert_equal "development", dbconsole[:environment]
+      assert_equal "custom", dbconsole.connection
+    end
+  end
+
+  def test_specifying_a_missing_connection
+    app_db_config({}) do
+      e = assert_raises(ActiveRecord::AdapterNotSpecified) do
+        Rails::Command.invoke(:dbconsole, ["-c", "i_do_not_exist"])
+      end
+
+      assert_includes e.message, "'i_do_not_exist' connection is not configured."
+    end
+  end
+
+  def test_specifying_a_missing_environment
+    app_db_config({}) do
+      e = assert_raises(ActiveRecord::AdapterNotSpecified) do
+        Rails::Command.invoke(:dbconsole)
+      end
+
+      assert_includes e.message, "'test' database is not configured."
+    end
+  end
+
   def test_print_help_short
     stdout = capture(:stdout) do
       Rails::Command.invoke(:dbconsole, ["-h"])